fix(review): move test_merged_hook_config_overlay_wins to tests/hooks/
The test exercises merge_hook_configs from autoskillit.hooks._hook_settings —
a pure hooks-layer utility with no server dependency. It belongs alongside
other _hook_settings tests in tests/hooks/, not in tests/server/.
